Oss release 26.0.2 created 2020-08-28-14-38
see CHANGELOG.txt for details

Original commit sha: c1dcea5ff0a8b5e6b6acc95e8578b6fe1e4ffe4d

+++ framework/Components/src/ResourceSerializationHelper.cpp
@@ -37,7 +37,7 @@ namespace ramses_internal
         {
             VoidOutputStream stream;
             SerializeResourceMetadata(stream, resource);
-            return stream.getSize();
+            return static_cast<uint32_t>(stream.getSize());
         }
 
         DeserializedResourceHeader ResourceFromMetadataStream(IInputStream& input)
